# dotnet/data-api/src/Smilr.DotNet/Models/Bulk.cs
/*
 * Smilr API
 *
 * Smilr microservice, RESTful data API
 *
 * OpenAPI spec version: 6.2.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;
using Smilr.DotNet.Attributes;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;

namespace Smilr.DotNet.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class Bulk : IEquatable<Bulk>
    { 
        /// <summary>
        /// Gets or Sets Events
        /// </summary>
        [Required]
        [DataMember(Name="events")]
        public List<Event> Events { get; set; }

        /// <summary>
        /// Gets or Sets Feedback
        /// </summary>
        [Required]
        [DataMember(Name="feedback")]
        public List<Feedback> Feedback { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Bulk {\n");
            sb.Append("  Events: ").Append(Events).Append("\n");
            sb.Append("  Feedback: ").Append(Feedback).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonSerializer.Serialize(this);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((Bulk)obj);
        }

        /// <summary>
        /// Returns true if Bulk instances are equal
        /// </summary>
        /// <param name="other">Instance of Bulk to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Bulk other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Events == other.Events ||
                    Events != null &&
                    Events.SequenceEqual(other.Events)
                ) && 
                (
                    Feedback == other.Feedback ||
                    Feedback != null &&
                    Feedback.SequenceEqual(other.Feedback)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Events != null)
                    hashCode = hashCode * 59 + Events.GetHashCode();
                    if (Feedback != null)
                    hashCode = hashCode * 59 + Feedback.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(Bulk left, Bulk right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(Bulk left, Bulk right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}